fix restic tests to work with different versions...
Update the check method to work with basenames, and also use the backup path directly
instead of guessing restore/output path, as the check function already does
fs.walk.

This issue mostly happend due to different restic versions usage of relative or absolute paths
when restoring snapshots.

Another solution is to use a fixed version of restic, but I hope this will work!